# Conflicts:
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/AcumulaciExpedientsApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/AvisosApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/DadesEspecifiquesApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/EstatsApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/ExpedientsApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/ExpedientsRelacionatsApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/Expedients_Api.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/PersonesInteressades_Api.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/PersonesSollicitudApi.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/PersonesSollicitud_Api.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/Persones_Api.java
#	gpaserveisRestApiClientExpedients/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaexpedients/SollicitudsApi.java
#	gpaserveisRestApiClientProcediments/src/main/java/es/bcn/gpa/gpaserveis/rest/client/api/gpaprocediments/ProcedimentsApi.java

+++ gpaserveisWeb/src/main/java/es/bcn/gpa/gpaserveis/business/impl/ExpedientsServiceImpl.java
@@ -212,7 +212,7 @@ public class ExpedientsServiceImpl implements ExpedientsService {
 		}
 
 		try {
-			SollicitudsRDTO sollicitudsRDTO = sollicitudsApi.consultarDadesSollicituds(id);
+			SollicitudsRDTO sollicitudsRDTO = sollicitudsApi.consultarDadesSollicituds(id.intValue());
 
 			if (log.isDebugEnabled()) {
 				log.debug("consultarDadesSollicitud(BigDecimal) - fi"); //$NON-NLS-1$
